Rename the project to htspy
