+++ src/htspy/_bamrecord.c
