+++ src/htspy/_bamrecord.pyi
