+++ src/htspy/bam.py
