+++ src/htspy/bgzf.py
