+++ tox.ini
@@ -7,7 +7,7 @@ deps=coverage
      pytest
 commands =
     # Create HTML coverage report for humans and xml coverage report for external services.
-    coverage run --source=pybam -m py.test -v tests
+    coverage run --source=htspy -m py.test -v tests
     coverage html
     coverage xml
 
@@ -17,7 +17,7 @@ deps=flake8
      mypy
 commands =
     flake8 src tests setup.py
-    mypy src/pybam tests/
+    mypy src/htspy tests/
 
 [testenv:twine_check]
 deps=build
